Cambio de clases y agregado  la evencia sin terminar

+++ src/analysis_unit/analysis_unit/services/analysis_unit.service.ts
@@ -3,12 +3,15 @@ import { AnalysisUnitDto } from '../dto/analysis_unit.dto';
 import { AnalysisUnit } from '../entity/analysis_unit';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
+import { CaseStudyContextAU } from 'src/case_study/case_study/entity/case_study_context_au';
 
 @Injectable()
 export class AnalysisUnitService {
   constructor(
     @InjectRepository(AnalysisUnit)
     private readonly analysisUnitRepository: Repository<AnalysisUnit>,
+    @InjectRepository(CaseStudyContextAU)
+    private readonly caseStudyContextRepository: Repository<CaseStudyContextAU>,
   ) {}
 
   async create(createAnalysisUnitDto: AnalysisUnitDto) {
@@ -34,15 +37,27 @@ export class AnalysisUnitService {
   async update(id: number, createAnalysisUnitDto: AnalysisUnitDto) {
     if (await this.findOne(id)) {
       await this.analysisUnitRepository.update(id, createAnalysisUnitDto);
-      const auUpdate = await this.findOne(id);
-      return auUpdate;
     }
   }
 
   async remove(id: number) {
+    if (
+      await this.caseStudyContextRepository.findOne({
+        where: { analysisUnit: id },
+      })
+    ) {
+      throw new NotFoundException(
+        `Can't delete analysis unit because it's assigned to a context`,
+      );
+    }
     const deleteAnalysisUnit = await this.analysisUnitRepository.delete(id);
     if (!deleteAnalysisUnit.affected) {
       throw new NotFoundException(`Analysis unit with id ${id} not found`);
     }
+  }
+
+  //Reporte
+  async findByEvidenceType(typeEvidence: number) {
+    return this.analysisUnitRepository.find({ where: { typeEvidence } });
   }
 }
